Create BaseAuthenticatedRequest and BaseResponse
