+++ lib/bloom_trade_client/requests/base_authenticated_request.rb
@@ -1,0 +1,14 @@
+module BloomTradeClient
+  class BaseAuthenticatedRequest
+
+    include APIClientBase::Request.module
+    attribute :token, String
+
+    def headers
+      {
+        "Authorization" => "Bearer #{token}",
+        "Content-Type" => "application/json",
+      }
+    end
+  end
+end
