+++ lib/bloom_trade_client/responses/base_response.rb
@@ -1,0 +1,16 @@
+module BloomTradeClient
+  class BaseResponse
+    include APIClientBase::Response.module
+
+    attribute :body, String, lazy: true, default: :default_body
+
+    private
+
+    def default_body
+      JSON.parse(raw_response.body).with_indifferent_access
+    end
+
+  end
+end
+
+
